requirements.txt
# Puget Sound OSINT - Dependencies

# Core
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
PyYAML>=6.0

# Async HTTP
aiohttp>=3.9.0

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Detection/ML
ultralytics>=8.0.0  # YOLOv8
torch>=2.0.0

# Utilities
python-dateutil>=2.8.0
requests>=2.31.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=23.0.0
ruff>=0.1.0
